add logic of product buying. regroup funcs
