+++ project_files/templates/admin/add_blocked.html
@@ -1,6 +1,6 @@
 
 
- <form action="/add-blocked" method='post'>
+ <form action="/add-blocked" method="post">
    <label for="username">username</label>
       <input type="text" name="username"/>
       <br>
