+++ project_files/templates/admin/add_product.html
@@ -1,6 +1,6 @@
 
 
- <form action="/add-product" method='post'>
+ <form action="/add-product" method="post">
    <label for="name">name</label>
       <input type="text" name="name"/>
       <br>
