+++ project_files/templates/admin/add_user.html
@@ -1,6 +1,6 @@
 
 
- <form action="/add-user" method='post'>
+ <form action="/add-user" method="post">
    <label for="username">username</label>
       <input type="text" name="username"/>
    <br>
