+++ project_files/templates/admin/update_blocked.html
@@ -1,6 +1,6 @@
 
 
- <form action="/update-blocked/{{blocked.id}}" method='post'>
+ <form action="/update-blocked/{{blocked.id}}" method="post">
 
    <label for="username">username:</label>
       <input type="text" name="username" value="{{blocked.username}}"/>
