+++ project_files/templates/admin/update_product.html
@@ -1,6 +1,6 @@
 
 
- <form action="/update-product/{{product.id}}" method='post'>
+ <form action="/update-product/{{product.id}}" method="post">
 
    <label for="name">name</label>
       <input type="text" name="name" value="{{product.name}}"/>
