+++ project_files/templates/admin/update_user.html
@@ -1,6 +1,6 @@
 
 
- <form action="/update-user/{{user.id}}" method='post'>
+ <form action="/update-user/{{user.id}}" method="post">
 
    <label for="username">username</label>
       <input type="text" name="username" value="{{user.username}}"/>
